pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "toolcli"
version = "0.1.0"
description = "Maintainable scaffold for an Ollama tool-calling CLI."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "rich>=13.7",
]

[project.scripts]
toolcli = "toolcli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
